requirements.txt
ultralytics
opencv-python-headless
requests
python-dotenv
